Analyse by years now that arithmetic is possible
This study was always planned to do analysis by age rather than year, to
make it just a tiny bit more interesting. But when we first wrote it
ehrQL didn't support arithmetic (how things have changed).

Note that this is backwards incompatible with the acceptance test in
Data Builder which checks the output of this study.